Move stats stuff to a separate package
