+++ src/main/java/fr/giovanni75/tarot/objects/Game.java
@@ -4,7 +4,8 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import fr.giovanni75.tarot.DateRecord;
-import fr.giovanni75.tarot.LocalStats;
+import fr.giovanni75.tarot.stats.GlobalStats;
+import fr.giovanni75.tarot.stats.LocalStats;
 import fr.giovanni75.tarot.Maps;
 import fr.giovanni75.tarot.Tarot;
 import fr.giovanni75.tarot.enums.*;
