+++ src/main/java/fr/giovanni75/tarot/objects/Player.java
@@ -2,7 +2,7 @@ package fr.giovanni75.tarot.objects;
 
 import com.google.gson.JsonObject;
 import fr.giovanni75.tarot.DateRecord;
-import fr.giovanni75.tarot.LocalStats;
+import fr.giovanni75.tarot.stats.LocalStats;
 import fr.giovanni75.tarot.Tarot;
 
 import java.util.HashMap;
